+++ lib/wipe_out/execution/execute_root_plan.rb
@@ -1,0 +1,15 @@
+module WipeOut
+  module Execution
+    class ExecuteRootPlan
+      method_object :plan, :record
+
+      def call
+        ExecuteAround.call(:around_all, [plan], plan.plugins) do
+          record.transaction do
+            Execute.call(plan, record, plan.plugins)
+          end
+        end
+      end
+    end
+  end
+end