+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v2.0.0
+
+Breaking changes:
+
+* Rename `build_root_plan` to `build_plan`
+* Rename `config` to `configure`- used only for configuration. To fetch config, old method applies
+* Rename `PluginBase` to `Plugin`
+* Execution by default will call `save!` instead of `save(validate: false)`
+
 # v1.1.1
 
 Fix wipe-out when relation is nil